Added auth routes, changed error reporting, and moved bible routes to proper place
