--- backend/middleware/auth.js
@@ -1,3 +1,0 @@
-// Copyright (c) 2019 WingGroup
-
-//USED FOR USER SESSION AUTHETICATION
