+++ backend/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "PORT=3001 nodemon index.js"
+    "start": "PORT=3001 node index.js"
   },
   "repository": {
     "type": "git",
@@ -19,7 +19,7 @@
     "body-parser": "^1.18.3",
     "express": "^4.16.4",
     "mongoose": "^5.4.16",
-    "nodemon": "^1.18.10",
+    "nodemon": "^1.18.11",
     "pug": "^2.0.3",
     "request": "^2.88.0"
   }
