+++ backend/routes/users.js
@@ -10,6 +10,9 @@ router.get('/', userController.list_users);
 // POST /users/
 router.post('/', userController.create_user);
 
+// GET /users/:id
+router.get('/:id', userController.get_user);
+
 // DELETE /users/:id
 router.delete('/:id', userController.delete_user);
